Update-version

+++ RGBA.Optio.Core/Entities/AbstractClass.cs
@@ -1,11 +1,10 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using System.Numerics;
 namespace Optio.Core.Entities
 {
     public abstract class AbstractClass
     {
         [Key]
-        public BigInteger Id { get; set; }
+        public long Id { get; set; }
      
     }
 }
