+++ RGBA.Optio.Core/Interfaces/IChannelRepo.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 
 namespace Optio.Core.Interfaces
 {
-    public interface IChannelRepo : ICrudRepo<Channels,BigInteger>
+    public interface IChannelRepo : ICrudRepo<Channels,long>
     {
         Task<IEnumerable<Channels>> GetAllActiveChannelAsync();
     }
