+++ RGBA.Optio.Core/Interfaces/ILocationRepo.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace Optio.Core.Interfaces
 {
-    public interface ILocationRepo:ICrudRepo<Location, BigInteger>
+    public interface ILocationRepo:ICrudRepo<Location, long>
     {
         Task<IEnumerable<Location>> GetAllActiveLocationAsync();
     }
