+++ RGBA.Optio.Core/Interfaces/ITransactionRepo.cs
@@ -3,11 +3,11 @@ using System.Numerics;
 
 namespace Optio.Core.Interfaces
 {
-    public interface ITransactionRepo:ICrudRepo<Transaction, BigInteger>
+    public interface ITransactionRepo:ICrudRepo<Transaction, long>
     {
         Task<IEnumerable<Transaction>> GetAllWithDetailsAsync();
 
-        Task<Transaction> GetByIdWithDetailsAsync(BigInteger Id);
+        Task<Transaction> GetByIdWithDetailsAsync(long Id);
 
         Task<IEnumerable<Transaction>> GetAllActiveAsync();
     }
