+++ RGBA.Optio.Domain/Interfaces/IMerchantRelatedService.cs
@@ -1,19 +1,13 @@
 ﻿using RGBA.Optio.Domain.Interfaces.InterfacesForTransaction;
 using RGBA.Optio.Domain.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Numerics;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RGBA.Optio.Domain.Interfaces
 {
     public interface IMerchantRelatedService:IAddInfo<MerchantModel>,IAddInfo<locationModel>,
-        IGetInfo<locationModel, BigInteger>,IGetInfo<MerchantModel,BigInteger>,
-        IRemoveInfo<locationModel, BigInteger>,IRemoveInfo<MerchantModel, BigInteger>,
-        IUpdateInfo<locationModel, BigInteger>,IUpdateInfo<MerchantModel, BigInteger>
+        IGetInfo<locationModel, long>,IGetInfo<MerchantModel,long>,
+        IRemoveInfo<locationModel, long>,IRemoveInfo<MerchantModel, long>,
+        IUpdateInfo<locationModel, long>,IUpdateInfo<MerchantModel, long>
     {
-        Task<bool> AssignLocationtoMerchant(BigInteger Merchantid, BigInteger Locationid);
+        Task<bool> AssignLocationtoMerchant(long Merchantid, long Locationid);
     }
 }
