+++ RGBA.Optio.Domain/Interfaces/ITransactionRelatedService.cs
@@ -1,13 +1,12 @@
 ﻿using RGBA.Optio.Domain.Interfaces.InterfacesForTransaction;
 using RGBA.Optio.Domain.Models;
-using System.Numerics;
 
 namespace RGBA.Optio.Domain.Interfaces
 {
     public interface ITransactionRelatedService:IAddInfo<ChanellModel>, IAddInfo<CategoryModel>, IAddInfo<TransactionTypeModel>,
-        IGetInfo<ChanellModel,BigInteger>, IGetInfo<CategoryModel, BigInteger>, IGetInfo<TransactionTypeModel, BigInteger>,
-        IRemoveInfo<ChanellModel, BigInteger>, IRemoveInfo<CategoryModel, BigInteger>, IRemoveInfo<TransactionTypeModel, BigInteger>,
-        IUpdateInfo<ChanellModel, BigInteger>, IUpdateInfo<CategoryModel, BigInteger>, IUpdateInfo<TransactionTypeModel, BigInteger>
+        IGetInfo<ChanellModel,long>, IGetInfo<CategoryModel, long>, IGetInfo<TransactionTypeModel, long>,
+        IRemoveInfo<ChanellModel, long>, IRemoveInfo<CategoryModel, long>, IRemoveInfo<TransactionTypeModel, long>,
+        IUpdateInfo<ChanellModel, long>, IUpdateInfo<CategoryModel, long>, IUpdateInfo<TransactionTypeModel, long>
     {
     }
 }
