+++ RGBA.Optio.Domain/Models/CategoryModel.cs
@@ -1,5 +1,4 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using System.Numerics;
 
 namespace RGBA.Optio.Domain.Models
 {
@@ -10,6 +9,7 @@ namespace RGBA.Optio.Domain.Models
         [Display(Name = "Transaction category name")]
         [RegularExpression(@"^[a-zA-Z\s]*$", ErrorMessage = "Transaction Type Name should contain only letters and spaces.")]
         public required string TransactionCategory { get; set; }
-        public required BigInteger TransactionTypeID { get; set; }
+
+        public required long TransactionTypeID { get; set; }
     }
 }
